Fasil/requirements.txt
asgiref==3.7.2
Authlib==1.2.1
azure-core==1.29.4
azure-identity==1.14.0
azure-storage-blob==12.18.2
backend==0.2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.2.5
django-logging-json==1.15
django-oauth-toolkit==2.3.0
django-storages==1.14.1
django-taggit==4.0.0
elastic-transport==8.4.1
elasticsearch==8.10.0
idna==3.4
isodate==0.6.1
jwcrypto==1.5.0
msal==1.24.1
msal-extensions==1.0.0
oauthlib==3.2.2
Pillow==10.0.1
portalocker==2.8.2
psycopg2-binary==2.9.8
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.21.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==6.6.0
typing_extensions==4.8.0
urllib3==1.26.16
whitenoise==6.5.0
wrapt==1.15.0
